forge fmt

+++ src/basics/2-votingv2.sol
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity >=0.7.0 <0.9.0;
 /// @title Voting with delegation.
-// Try Possible Improvements Possible Improvements 
+// Try Possible Improvements Possible Improvements
 // https://docs.soliditylang.org/en/v0.8.28/solidity-by-example.html#possible-improvements
-// Currently, many transactions are needed to assign the rights to vote to all participants. 
+// Currently, many transactions are needed to assign the rights to vote to all participants.
 // Moreover, if two or more proposals have the same number of votes, winningProposal() is not able to register a tie.
 // Can you think of a way to fix these issues?
 
@@ -12,16 +12,16 @@ contract BallotV2 {
     // be used for variables later.
     // It will represent a single voter.
     struct Voter {
-        uint weight; // weight is accumulated by delegation
-        bool voted;  // if true, that person already voted
+        uint256 weight; // weight is accumulated by delegation
+        bool voted; // if true, that person already voted
         address delegate; // person delegated to
-        uint vote;   // index of the voted proposal
+        uint256 vote; // index of the voted proposal
     }
 
     // This is a type for a single proposal.
     struct Proposal {
-        bytes32 name;   // short name (up to 32 bytes)
-        uint voteCount; // number of accumulated votes
+        bytes32 name; // short name (up to 32 bytes)
+        uint256 voteCount; // number of accumulated votes
     }
 
     address public chairperson;
@@ -41,14 +41,11 @@ contract BallotV2 {
         // For each of the provided proposal names,
         // create a new proposal object and add it
         // to the end of the array.
-        for (uint i = 0; i < proposalNames.length; i++) {
+        for (uint256 i = 0; i < proposalNames.length; i++) {
             // `Proposal({...})` creates a temporary
             // Proposal object and `proposals.push(...)`
             // appends it to the end of `proposals`.
-            proposals.push(Proposal({
-                name: proposalNames[i],
-                voteCount: 0
-            }));
+            proposals.push(Proposal({name: proposalNames[i], voteCount: 0}));
         }
     }
 
@@ -65,14 +62,8 @@ contract BallotV2 {
         // functions are called correctly.
         // As a second argument, you can also provide an
         // explanation about what went wrong.
-        require(
-            msg.sender == chairperson,
-            "Only chairperson can give right to vote."
-        );
-        require(
-            !voters[voter].voted,
-            "The voter already voted."
-        );
+        require(msg.sender == chairperson, "Only chairperson can give right to vote.");
+        require(!voters[voter].voted, "The voter already voted.");
         require(voters[voter].weight == 0);
         voters[voter].weight = 1;
     }
@@ -124,7 +115,7 @@ contract BallotV2 {
 
     /// Give your vote (including votes delegated to you)
     /// to proposal `proposals[proposal].name`.
-    function vote(uint proposal) external {
+    function vote(uint256 proposal) external {
         Voter storage sender = voters[msg.sender];
         require(sender.weight != 0, "Has no right to vote");
         require(!sender.voted, "Already voted.");
@@ -139,11 +130,9 @@ contract BallotV2 {
 
     /// @dev Computes the winning proposal taking all
     /// previous votes into account.
-    function winningProposal() public view
-            returns (uint winningProposal_)
-    {
-        uint winningVoteCount = 0;
-        for (uint p = 0; p < proposals.length; p++) {
+    function winningProposal() public view returns (uint256 winningProposal_) {
+        uint256 winningVoteCount = 0;
+        for (uint256 p = 0; p < proposals.length; p++) {
             if (proposals[p].voteCount > winningVoteCount) {
                 winningVoteCount = proposals[p].voteCount;
                 winningProposal_ = p;
@@ -154,9 +143,7 @@ contract BallotV2 {
     // Calls winningProposal() function to get the index
     // of the winner contained in the proposals array and then
     // returns the name of the winner
-    function winnerName() external view
-            returns (bytes32 winnerName_)
-    {
+    function winnerName() external view returns (bytes32 winnerName_) {
         winnerName_ = proposals[winningProposal()].name;
     }
 }
